requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
pillow==10.4.0
qrcode[pil]==8.0
jinja2==3.1.4
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
python-dotenv==1.0.1
